+++ tests/test_repo.py
@@ -34,3 +34,15 @@ def test_init_commit_repo_with_wrong_structure(tmpdir):  # noqa
     os.mkdir(project)
     with pytest.raises(RuntimeError):
         repo.init_commit_repo(project, struct)
+
+
+def test_add_tag(tmpdir):  # noqa
+    project = "my_project"
+    struct = {project: {
+        "my_file": "Some other content",
+        "my_dir": {"my_file": "Some more content"}}
+    }
+    structure.create_structure(struct)
+    repo.init_commit_repo(project, struct)
+    repo.add_tag(project, "v0.0")
+    repo.add_tag(project, "v0.1", "Message with whitespace")